+++ test/test_util.py
@@ -1,5 +1,5 @@
 import subprocess
-from test_utils import *
+from utils import *
 
 
 @all_files_in_dir('util_clk_reset')
